// code/plugins/mdb-angular-calendar/src/lib/calendar.options.interface.ts
export interface MdbCalendarOptions {
  weekdays: string[];
  months: string[];
  monthsShort: string[];
  todayCaption: string;
  monthCaption: string;
  weekCaption: string;
  listCaption: string;
  allDayCaption: string;
  noEventsCaption: string;
  summaryCaption: string;
  descriptionCaption: string;
  startCaption: string;
  endCaption: string;
  addCaption: string;
  deleteCaption: string;
  editCaption: string;
  closeCaption: string;
  addEventModalCaption: string;
  editEventModalCaption: string;
}